+++ src/main.tsx
@@ -5,7 +5,7 @@ import { Global } from '@emotion/react';
 import globalStyles from './styles/global.styles';
 import './index.css';
 import { ToastProvider } from './context/ToastContext';
-import { CartListProvider } from './context/useCartListContext';
+import { CartListProvider } from './context/CartListContext';
 
 async function enableMocking() {
   const { worker } = await import('./mocks/browser');
